+++ packages/react/src/components/SuggestionMenu/getDefaultReactSlashMenuItems.tsx
@@ -25,9 +25,11 @@ import {
   RiText,
   RiVolumeUpFill,
 } from "react-icons/ri";
+
+import { IconType } from "react-icons";
 import { DefaultReactSuggestionItem } from "./types.js";
 
-const icons = {
+const icons: Record<string, IconType> = {
   heading: RiH1,
   heading_2: RiH2,
   heading_3: RiH3,